+++ src/main/java/nextstep/mission/view/ResultViewer.java
@@ -1,4 +1,9 @@
-package nextstep.mission;
+package nextstep.mission.view;
+
+import nextstep.mission.domain.Participant;
+import nextstep.mission.domain.Ladder;
+import nextstep.mission.domain.LadderReader;
+import nextstep.mission.domain.Line;
 
 import java.util.List;
 import java.util.stream.Collectors;
