+++ src/test/java/nextstep/mission/domain/LadderTest.java
@@ -1,5 +1,8 @@
-package nextstep.mission;
+package nextstep.mission.domain;
 
+import nextstep.mission.domain.Ladder;
+import nextstep.mission.domain.Participant;
+import nextstep.mission.domain.Participants;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
