+++ src/test/java/nextstep/mission/domain/LineTest.java
@@ -1,9 +1,9 @@
-package nextstep.mission;
+package nextstep.mission.domain;
 
+import nextstep.mission.domain.Line;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
